fix(ME-103): Refactory code Deleter flow whit the find

+++ src/Backoffice/Security/RolePermissions/Application/Delete/RolePermissionDeleter.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace MedineTech\Backoffice\Security\RolePermissions\Application\Delete;
 
 use MedineTech\Backoffice\Security\RolePermissions\Domain\RolePermission;
+use MedineTech\Backoffice\Security\RolePermissions\Domain\RolePermissionNotFoundException;
 use MedineTech\Backoffice\Security\RolePermissions\Domain\RolePermissionRepository;
 
 final class RolePermissionDeleter
@@ -14,13 +15,20 @@ final class RolePermissionDeleter
     ) {
     }
 
-    public function __invoke(RolePermissionDeleterRequest $request): void
+    public function __invoke(RolePermissionDeleterRequest $request): RolePermissionDeleterResponse
     {
-        $rolePermission = RolePermission::find(
-            $request->roleId(),
-            $request->permissionId()
-        );
+
+        $rolePermission = $this->repository->find($request->roleId(), $request->permissionId());
+
+        if (null === $rolePermission) {
+            throw new RolePermissionNotFoundException($request->roleId(), $request->permissionId());
+        }
 
         $this->repository->delete($rolePermission);
+
+        return new RolePermissionDeleterResponse(
+            $rolePermission->roleId(),
+            $rolePermission->permissionId()
+        );
     }
 }
